Prevent infinite loop in worker resulting from the number of workers becoming larger than the number of jobs
This change addresses a bug where Oxidized's worker loop will deadlock if the number of workers becomes larger than the number of nodes. This behavior may be triggered rather easily via the use if the /next API endpoint when next_adds_job option is enabled and the node list is small (100% probability of occurrence with 1 node).